refactor history to drink log
